Bugfix on countrycode.py::countrycode
Discovered that when codes argument is a pandas Series, the countrycode function
treats the argument as a shallow copy.  When modified (e.g. replacing non-matched
country names with NaN), these changes were reflected in the original object as
well.  Solution is to make a deep copy of the data in the codes object.